Fixed parameters tuple access with unicode index
On executing context for a prepared statement which calls parameterized
functions, stored procedures, etc. HANA returns parametermetadata
containing the parameters name as id. On execution cursor.py used this
id for index access to the parameters tuple. Now an enumeration is used
to get the position of the current parameter/parametermetadata and for
access to the paramters tuple.

Executing a prepared statement can return parametermetadata which led
to "unknwon part kind" errors. The parameterdataset is now recognized
as valid part kind (but is ignored).