CESU-8: Fixed various decoding problems
The CESU-8 decoder wasn't suited to decode input that doesn't
indicate CESU-8 in the first byte. Now the decoder performs chunked
decoding what means it decodes everything until the first CESU-8
surrogate pair with UTF-8. Afterwards it verifies that the CESU-8
surrogate pair is within the permitted range otherwise it is probably
a 3 byte UTF-8 sequence and falls back to the UTF-8 decoder.
The CESU-8 decoder repeats this algorithm until it decoded the
full input or an error occurs.