Fix UnicodeDecodeError
Issue:
When there are some data like emoji which can not decode by 'utf-8', it could cause such error:

~\AppData\Local\Continuum\anaconda3\envs\work\lib\site-packages\pyhdb\protocol\lobs.py in _decode_lob_data(cls, payload_data)
     77     @classmethod
     78     def _decode_lob_data(cls, payload_data):
---> 79         return payload_data.decode(cls.encoding) if cls.encoding else payload_data
     80 
     81     def __init__(self, init_value='', lob_header=None, connection=None):

UnicodeDecodeError: 'utf-8' codec can't decode byte 0xed in position 37: invalid continuation byte

The problem here is the previous function assumes data is pure bytes type encoding from str, however it may happen that the data is not clean. 

Solution:
If the data is not clean or can not decode to str, then return it empty with same length (or better solution? remove it?)